use (unsigned?) longs for lengths

+++ src/audio/datasource_filereader.h
@@ -13,7 +13,7 @@ class FileReaderDataSource: public IDataSource
 		void reset();
 		bool exhausted();
 		long getpos();
-		int read(char* const buffer, int len);
+		int read(char* const buffer, unsigned long len);
 	private:
 		FILE* FileHandle;
 };
